Build: Fixes to allow compiling the whole solution, rather than just certain targets.
* TestPAL: Compile failure due to an MSVC warning about assigning 32-bit shift result to 64-bit variable. Also added to CIS.
* UE4EditorServices: Conflict in variable name for hInstance. Disabled building this target by default for Windows and Linux, since it only usefully does anything on Mac.

#codereview Dmitry.Rekman, Michael.Trepka
#rb none
#lockdown Nick.Penwarden

[CL 3188474 by Ben Marsh in Main branch]

+++ Engine/Source/Programs/TestPAL/Private/Main.cpp
@@ -458,7 +458,7 @@ int32 GetAllocationSizeTest(const TCHAR* CommandLine)
 	{
 		for (SIZE_T AlignmentPower = 4; AlignmentPower <= 7; ++AlignmentPower)
 		{
-			SIZE_T Alignment = (1 << AlignmentPower);
+			SIZE_T Alignment = ((SIZE_T)1 << AlignmentPower);
 
 			Allocation New;
 			New.RequestedSize = Size;
